Add files via upload
Developed To-Do Application using Python Programming language for the script and Tkinter for the Windows Application. The main purpose of this Application is to add the list of tasks by the user and ones they completed and the user can delete them by entering the serial number of that particular task in the text box and click the button of delete and then we can exit.